+++ src/rust/src/lib.rs
@@ -7,6 +7,7 @@ mod migration;
 mod column;
 mod bincount;
 mod sir;
+mod vitals;
 pub use laser_frame::LaserFrame;
 pub use distributions::Distribution;
 pub use column::Column;
@@ -20,4 +21,5 @@ extendr_module! {
     use column;
     use bincount;
     use sir;
+    use vitals;
 }